+++ R/ngram.R
@@ -1,0 +1,151 @@
+#' Generate ngrams from tokenlist
+#'
+#' `step_ngram` creates a *specification* of a recipe step that
+#'  will convert a list of tokens into a list of ngram of tokens.
+#'
+#' @param recipe A recipe object. The step will be added to the
+#'  sequence of operations for this recipe.
+#' @param ... One or more selector functions to choose variables.
+#'  For `step_ngram`, this indicates the variables to be encoded
+#'  into a list column. See [recipes::selections()] for more
+#'  details. For the `tidy` method, these are not currently used.
+#' @param role Not used by this step since no new variables are
+#'  created.
+#' @param columns A list of tibble results that define the
+#'  encoding. This is `NULL` until the step is trained by
+#'  [recipes::prep.recipe()].
+#' @param n The number of words in the n-gram. This must be an integer greater 
+#' than or equal to 1. Defaults to 3.
+#' @param delim The separator between words in an n-gram. Defaults to "_".
+#' @param skip A logical. Should the step be skipped when the
+#'  recipe is baked by [recipes::bake.recipe()]? While all
+#'  operations are baked when [recipes::prep.recipe()] is run, some
+#'  operations may not be able to be conducted on new data (e.g.
+#'  processing the outcome variable(s)). Care should be taken when
+#'  using `skip = TRUE` as it may affect the computations for
+#'  subsequent operations.
+#' @param id A character string that is unique to this step to identify it.
+#' @param trained A logical to indicate if the recipe has been
+#'  baked.
+#' @return An updated version of `recipe` with the new step added
+#'  to the sequence of existing steps (if any).
+#' @examples
+#' library(recipes)
+#' 
+#' data(okc_text)
+#' 
+#' okc_rec <- recipe(~ ., data = okc_text) %>%
+#'   step_tokenize(essay0) %>%
+#'   step_ngram(essay0)
+#'   
+#' okc_obj <- okc_rec %>%
+#'   prep(training = okc_text, retain = TRUE)
+#' 
+#' juice(okc_obj, essay0) %>% 
+#'   slice(1:2)
+#' 
+#' juice(okc_obj) %>% 
+#'   slice(2) %>% 
+#'   pull(essay0) 
+#'   
+#' tidy(okc_rec, number = 2)
+#' tidy(okc_obj, number = 2)
+#' @export
+#' @seealso [step_stopwords()] [step_tokenfilter()] [step_tokenize()]
+step_ngram <-
+  function(recipe,
+           ...,
+           role = NA,
+           trained = FALSE,
+           columns = NULL,
+           n = 3L,
+           delim = "_",
+           skip = FALSE,
+           id = rand_id("ngram")
+  ) {
+    add_step(
+      recipe,
+      step_ngram_new(
+        terms = ellipse_check(...),
+        role = role,
+        trained = trained,
+        n = n,
+        delim = delim,
+        columns = columns,
+        skip = skip,
+        id = id
+      )
+    )
+  }
+
+step_ngram_new <-
+  function(terms, role, trained, columns, n, delim, skip, id) {
+    step(
+      subclass = "ngram",
+      terms = terms,
+      role = role,
+      trained = trained,
+      columns = columns,
+      n = n,
+      delim = delim,
+      skip = skip,
+      id = id
+    )
+  }
+
+#' @export
+prep.step_ngram <- function(x, training, info = NULL, ...) {
+  col_names <- terms_select(x$terms, info = info)
+  
+  check_list(training[, col_names])
+  
+  step_ngram_new(
+    terms = x$terms,
+    role = x$role,
+    trained = TRUE,
+    columns = col_names,
+    n = x$n,
+    delim = x$delim,
+    skip = x$skip,
+    id = x$id
+  )
+}
+
+#' @export
+bake.step_ngram <- function(object, new_data, ...) {
+  col_names <- object$columns
+  # for backward compat
+  
+  for (i in seq_along(col_names)) {
+    ngrammed_tokenlist <- tokenlist_ngram(new_data[, col_names[i], drop = TRUE],
+                                          n = object$n,
+                                          delim = object$delim)
+    
+    new_data[, col_names[i]] <- tibble(ngrammed_tokenlist)
+  }
+  new_data <- factor_to_text(new_data, col_names)
+  as_tibble(new_data)
+}
+
+#' @export
+print.step_ngram <-
+  function(x, width = max(20, options()$width - 30), ...) {
+    cat("ngramming for ", sep = "")
+    printer(x$columns, x$terms, x$trained, width = width)
+    invisible(x)
+  }
+
+#' @rdname step_ngram
+#' @param x A `step_ngram` object.
+#' @export
+tidy.step_ngram <- function(x, ...) {
+  if (is_trained(x)) {
+    res <- tibble(terms = x$terms)
+  } else {
+    term_names <- sel2char(x$terms)
+    res <- tibble(terms = term_names,
+                  value = na_chr)
+  }
+  res$id <- x$id
+  res
+}