+++ NEWS.md
@@ -4,6 +4,7 @@
 * `spacyr` have been added as an engine to `step_tokenize()`.
 * `step_lemma()` has been added to extract lemma attribute from tokenlists.
 * `step_pos_filter()` has been added to allow filtering of tokens bases on their pat of speech tags.
+* `step_ngram()` has been added to generate ngrams from tokenlists.
 
 # textrecipes 0.1.0